Agentes/reciclador/pyproject.toml
[project]
name = "reciclador"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.18.0",
    "mirascope[google]>=1.25.6",
    "mysql-connector-python>=9.5.0",
]
